added all Entities, Services, Repositories

+++ lms/src/main/java/com/harman/lms/repository/BookRepository.java
@@ -3,4 +3,4 @@ package com.harman.lms.repository;
 import com.harman.lms.entity.Book;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookRepository extends JpaRepository<Book, Integer> {}
+public interface BookRepository extends JpaRepository<Book, Long> {}
